rename Models::Diary to DiaryContainer
