+++ lib/tdiary.rb
@@ -50,7 +50,7 @@ module TDiary
 	# Diary model class
 	autoload :Style,                    'tdiary/style'
 	autoload :Comment,                  'tdiary/comment'
-	autoload :Models,                   'tdiary/models'
+	autoload :DiaryContainer,           'tdiary/diary_container'
 
 	# Routing and Dispatch
 	autoload :Dispatcher,               'tdiary/dispatcher'
